#29 - Refactor event handling between calendar and home components
Moved event creation and viewing logic from CalendarioComponent to HomeComponent by emitting addEvent and viewEvent outputs. Updated home.html and home.ts to handle these events. Cleaned up calendario.module.ts imports and removed unused declarations.
