+++ src/app/home/home.html
@@ -4,4 +4,6 @@
   [modalData]="modalData"
   (close)="closeEventModal()"
   (edit)="editEvent($event)"
+  (addEvent)="onAddEvent($event)"
+  (viewEvent)="onViewEvent($event)"
 ></app-event-modal>
